feat(doc-data): Doc-string and Data-Tables   * Cleaned up test feature files   * Added tests for parsing doc-strings and data-tables
