+++ src/__mocks__/features/index.js
@@ -1,0 +1,8 @@
+export * from './backgroundFeature'
+export * from './basicFeature'
+export * from './commentsFeature'
+export * from './docAndDataFeature'
+export * from './failStepFeature'
+export * from './parsedFeature'
+export * from './promiseFeature'
+export * from './tagsFeature'
