+++ src/__mocks__/features/promiseFeature.js
@@ -1,0 +1,6 @@
+export const promiseFeature = `@promises
+Feature: Test Promises
+  Scenario: Test definitions with promises 
+    Given I wait for 1000
+    Then the world test method should be called
+`
