+++ src/__mocks__/index.js
@@ -1,4 +1,4 @@
 export * from './definition'
-export * from './feature'
+export * from './features'
 export * from './steps'
 export * from './hooks'
